fix: low worker-shutdown-timeout for ingress-nginx
On upgrades, the old pod is shut down before the new one starts (for port binding reasons). The shutdown waits a long time, 240s by default, for connections to gracefully terminate - which doesn't happen because of open websockets, usually on the demo. This sets that limit low to kill those connections instead.